Directfs implementation.
Directfs is a fork of the gofer client. It basically merges the gofer client
and runsc fsgofer functionality. Directfs access host filesystem directly,
without going through the gofer server. It works on a donated FD to the root of
the mount. All other relevant mount options supported in gofer client are also
supported here.

PiperOrigin-RevId: 474738988